requirements.txt
python-telegram-bot>=21.7,<22.0
pymumble==1.6.1
